Flesh out `idscope` solution
Refine the explanation of IDREF scoping and related concepts, including a narrower definition of the problem and proposed solutions. Add details on the `idscope` attribute and its implications for ID resolution in HTML documents.

This PR is meant as a discussion point, to allow reasoning on this solution and the pain points it tries to address.